Refractor code into separate modules
